safari-534.53-branch/LayoutTests/canvas/philip/tests/2d.canvas.readonly.html

<!DOCTYPE html>
<title>Canvas test: 2d.canvas.readonly</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var c = document.createElement('canvas');
var d = ctx.canvas;
_assertDifferent(c, d, "c", "d");
try { ctx.canvas = c; } catch (e) {} // TODO: not sure whether this should throw or not...
_assertSame(ctx.canvas, d, "ctx.canvas", "d");


});
</script>
