LayoutTests/canvas/philip/tests/2d.canvas.readonly-expected.txt

< [index] >
2d.canvas.readonly
CanvasRenderingContext2D.canvas is readonly
References: 2d.canvas.attribute
Actual output:
Passed
